add key to ssh-agent in Travis

+++ deploy.sh
@@ -13,27 +13,37 @@ fi
 
 REPO=`git config remote.origin.url`
 SSH_REPO=${REPO/https:\/\/github.com\//git@github.com:}
+echo $REPO
+echo $SSH_REPO
+if [ ! -f $TRAVIS_BUILD_DIR/travis_key ]; then
+    echo "Missing Travis Key"
+    exit 0
+fi
+
+chmod 600 travis_key
+eval `ssh-agent -s`
+ssh-add travis_key
 
 if [ "$TRAVIS_BRANCH" = "master" ] && \
    [ "(ls -A $TRAVIS_BUILD_DIR/docs/html)" ]; then
-    git clone --branch=gh-pages $SSH_REPO gh-pages
+    git clone --branch=gh-pages $REPO gh-pages
     cd gh-pages
-    rm -r css favicon.png fonts index.html interface \
+    rm -rf css favicon.png fonts index.html interface \
        js lists media module page proc program search.html \
        sourcefile src tipuesearch type
     cp -r "$TRAVIS_BUILD_DIR"/docs/html/* .
     git add -A
     git commit -m "Development Documentation updated by travis job $TRAVIS_JOB_NUMBER for commits $TRAVIS_COMMIT_RANGE"
-    git push origin gh-pages
+    git push $SSH_REPO gh-pages
 fi
 
 if [[ $TRAVIS_BRANCH == release-* ]]; then
-    git clone --branch=gh-pages $SSH_REPO gh-pages
+    git clone --branch=gh-pages $REPO gh-pages
     cd gh-pages
     rm -rf $TRAVIS_BRANCH
     mkdir $TRAVIS_BRANCH
     cp -r "$TRAVIS_BUILD_DIR"/docs/html/* $TRAVIS_BRANCH
     git add -A
     git commit -m "$TRAVIS_BRANCH documentation updated by travis job $TRAVIS_JOB_NUMBER for commits $TRAVIS_COMMIT_RANGE"
-    git push origin gh-pages
+    git push $SSH_REPO gh-pages
 fi
